requirements.txt
fastapi==0.125.0
sqlalchemy==2.0.46
pydantic==2.12.5
PyJWT==2.11.0
asyncpg==0.31.0
uvicorn==0.35.0
alembic==1.16.5
